refactor: move FileChooseAction into one class
FileChooseAction was an inner class of BrowserFragment. Now move it to
standalone class to make BrowserFragment being thin.
